refactor: code cleanup, and Discord RPC improvements
Code cleanup:
- Add CPI_OutputModule_InitDefaults() helper and apply to all 5 output
  backends (Wave, DirectSound, WASAPI, FAudio, File) to eliminate
  boilerplate field assignments
- Add CP_InStream_Init() helper and apply to LocalFile and Internet
  stream setup
- Move GetFilePointer macro from CPI_Stream_LocalFile.c to globals.h
- Standardize codec allocation to CALLOC_TYPE in WAV and AAC codecs
- Standardize WAV Uninitialise error handling to null guard pattern

Compiler warning fixes:
- Wrap FLAC__NO_DLL and TAGLIB_STATIC in #ifndef guards to fix
  -Wmacro-redefined warnings
- Change m_iOpenDevice_Freq_Hz from int to UINT in CPI_Player_Engine.h
  to fix -Wsign-compare warnings
- Add (const BYTE*) cast in CPI_Player_FileAssoc.c for -Wpointer-sign
- Add (unsigned int*) casts for CLV column order calls in
  CPI_PlaylistWindow.c
- Fix DlgProc_Properties declaration to use INT_PTR CALLBACK to
  eliminate -Wcast-function-type-mismatch
- Fix bitrate comparison in CPI_Player_CoDec_MPEG.c to use matching
  unsigned types

Discord RPC:
- Move option() declarations before project() in CMakeLists.txt so
  ENABLE_DISCORD_RPC and ENABLE_FFMPEG are defined when the vcpkg
  feature selection block runs on first configure
- Add discord-rpc to default-features in vcpkg.json
- Default discord_rpc_enabled to 1 in profile.c
- Push current playback state to Discord immediately when the user
  enables RPC at runtime via the Options dialog
- Wrap Discord RPC and DSP debug log helpers in #ifdef _DEBUG

+++ CPI_Player_CoDec_MPEG.c
@@ -738,9 +738,9 @@ BOOL CPP_OMMP3_GetPCMBlock(CPs_CoDecModule* pModule, void *block, DWORD *size)
 		
 		bitrate = context->rate / context->frames;
 		
-		if (context->info.m_iBitRate_Kbs != (int)bitrate)
+		if (context->info.m_iBitRate_Kbs != bitrate)
 		{
-			context->info.m_iBitRate_Kbs = (int)bitrate;
+			context->info.m_iBitRate_Kbs = bitrate;
 		}
 	}
